// src/app/interfaces/producto.interface.ts
export interface Producto {
  id?: string;
  codigo: string;
  descripcion: string;
  cantidad: number | null;
  precio: number | null;
  cantidadDeseada?: number; // Añadimos esta propiedad opcional
}

export interface Producto {
  codigo: string;
  descripcion: string;
  cantidad: number | null;
  precio: number | null;
  cantidadDeseada?: number; // Añadimos esta propiedad opcional
}